git-svn-id: https://surf.svn.sourceforge.net/svnroot/surf/trunk@153 5f8b00c5-0c0e-0410-9123-c6bd0749f4b9

+++ surf/src/float_buffer.cc
@@ -153,7 +153,7 @@ float float_buffer::GetAbsFieldModuloHeight( int x, int y )
 void float_buffer::CorrectGamma (double gamma)
 {
 	for( int pos = 0; pos < n; pos++ )
-		b[pos] = pow( b[pos], gamma );
+		b[pos] = pow((double)b[pos], gamma);
 }
 
 // ----------------------------------------------------------------------------
